+++ CHANGELOG.md
@@ -110,6 +110,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove EOF character `0x1A` in shipping files
 - Rename `B. do Nordeste` to `Banco do Nordeste`
 - BankBillet views
+- Rename `formated` to `formatted`
 
 ### Security